requirements.txt
discord.py==2.1.0
openai==0.27.2
python-dotenv==0.21.0
requests==2.25.1
pyrandmeme==0.0.5
pyjokes==0.6.0